crisissynai/backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-generativeai
networkx
python-dotenv
httpx
